add controlpanel component
